+++ History.md
@@ -1,6 +1,7 @@
 0.3.1 / 2013-??-??
 ==================
   * Add function data to `jscoverage_report(dir)` function call (https://github.com/tntim96/JSCover/issues/69)
+  * Merging Empty Branches Causes IllegalStateException (https://github.com/tntim96/JSCover/issues/68)
   * Internal: Use Maven for dependency management (https://github.com/tntim96/JSCover/issues/9)
 
 0.3.0 / 2013-05-28